refactor:#1 로직 / UI 분리
