+++ src/app/items/components/item-grid.tsx
@@ -13,67 +13,15 @@ import {
 } from "@/components/ui/tooltip";
 import { TooltipTrigger } from "@radix-ui/react-tooltip";
 import { Coins, Info } from "lucide-react";
+import { ItemGridProps } from "../types";
+import { categoryFilter } from "../utils/filterItems";
+import { fetchItemList } from "../utils/fetchItemList";
+import Image from "next/image";
+import { removeHtmlTags } from "../utils/removeHtmlTags";
 
-interface ItemGridProps {
-  category: string;
-  items: ApiItem[];
-}
-
-export interface ApiItem {
-  id: string;
-  name: string;
-  description: string;
-  colloq: string;
-  plaintext: string;
-  from?: string[];
-  into?: string[];
-  image: {
-    full: string;
-    sprite: string;
-    group: string;
-    x: number;
-    y: number;
-    w: number;
-    h: number;
-  };
-  gold: {
-    base: number;
-    purchasable: boolean;
-    total: number;
-    sell: number;
-  };
-  tags: string[];
-  maps: Record<string, boolean>;
-  stats: Record<string, number>;
-  depth?: number;
-}
-
-function stripHtmlTags(html: string): string {
-  return html.replace(/<[^>]*>/g, "");
-}
-
-export function ItemGrid({ category, items }: ItemGridProps) {
-  const filteredItems = items.filter((item) => {
-    if (category === "all") return true;
-    if (category === "boots") return item.tags.includes("Boots");
-    if (category === "attack")
-      return item.tags.some((tag) =>
-        ["Damage", "CriticalStrike", "AttackSpeed"].includes(tag)
-      );
-    if (category === "magic")
-      return item.tags.some((tag) =>
-        ["SpellDamage", "Mana", "ManaRegen"].includes(tag)
-      );
-    if (category === "defense")
-      return item.tags.some((tag) =>
-        ["Health", "Armor", "SpellBlock", "HealthRegen"].includes(tag)
-      );
-    if (category === "support")
-      return item.tags.some((tag) =>
-        ["Active", "GoldPer", "Vision"].includes(tag)
-      );
-    return false;
-  });
+export async function ItemGrid({ category, items }: ItemGridProps) {
+  const { filteredItems } = categoryFilter(category, items);
+  const { version } = await fetchItemList();
 
   return (
     <div className="grid grid-cols-1 sm:grid-cols-2 md:grid-cols-3 lg:grid-cols-4 gap-4">
@@ -83,10 +31,12 @@ export function ItemGrid({ category, items }: ItemGridProps) {
             <div className="flex items-start justify-between">
               <div className="flex items-center space-x-3">
                 <div className="w-12 h-12 rounded bg-secondary/50 flex items-center justify-center overflow-hidden">
-                  <img
-                    src={`https://ddragon.leagueoflegends.com/cdn/14.6.1/img/item/${item.image.full}`}
+                  <Image
+                    src={`https://ddragon.leagueoflegends.com/cdn/${version}/img/item/${item.image.full}`}
                     alt={item.name}
                     className="w-full h-full object-cover"
+                    width={48}
+                    height={48}
                   />
                 </div>
                 <div>
@@ -106,7 +56,7 @@ export function ItemGrid({ category, items }: ItemGridProps) {
                   </TooltipTrigger>
                   <TooltipContent>
                     <p className="max-w-xs">
-                      {stripHtmlTags(item.description)}
+                      {removeHtmlTags(item.description)}
                     </p>
                   </TooltipContent>
                 </Tooltip>
@@ -128,6 +78,10 @@ export function ItemGrid({ category, items }: ItemGridProps) {
                   statName = `+${value} 마법 저항력`;
                 else if (key === "FlatHPPoolMod") statName = `+${value} 체력`;
                 else if (key === "FlatMPPoolMod") statName = `+${value} 마나`;
+                else if (key === "FlatHPRegenMod")
+                  statName = `+${value * 100}% 체력 재생`;
+                else if (key === "FlatMPRegenMod")
+                  statName = `+${value * 100}% 마나 재생`;
                 else if (key === "PercentAttackSpeedMod")
                   statName = `+${value * 100}% 공격 속도`;
                 else if (key === "PercentLifeStealMod")
@@ -150,7 +104,7 @@ export function ItemGrid({ category, items }: ItemGridProps) {
               })}
             </div>
             <CardDescription className="mt-3 line-clamp-3 text-xs">
-              {item.plaintext || stripHtmlTags(item.description)}
+              {item.plaintext || removeHtmlTags(item.description)}
             </CardDescription>
           </CardContent>
         </Card>
