+++ src/app/items/utils/fetchItemList.ts
@@ -1,0 +1,18 @@
+import { fetchLatestVersion } from "@/app/api/action";
+import { ApiItem, ItemPageLogicResult } from "../types";
+
+export async function fetchItemList(): Promise<ItemPageLogicResult> {
+  const version: string = await fetchLatestVersion();
+
+  const res = await fetch(
+    `https://ddragon.leagueoflegends.com/cdn/${version}/data/ko_KR/item.json`,
+    { cache: "force-cache" }
+  );
+  const { data } = await res.json();
+  const items = Object.entries<ApiItem>(data).map(([id, item]) => ({
+    ...item,
+    id,
+  }));
+
+  return { version, items };
+}
