+++ src/app/items/utils/removeHtmlTags.ts
@@ -1,0 +1,3 @@
+export function removeHtmlTags(desc: string): string {
+  return desc.replace(/<[^>]*>/g, "");
+}
