+++ src/app/layout.tsx
@@ -1,8 +1,7 @@
 import type React from "react";
 import type { Metadata } from "next";
-import "./globals.css";
-import Header from "@/components/header";
-import { ThemeProvider } from "@/components/theme-provider";
+import "../styles/globals.css";
+import Header from "@/app/components/header";
 
 export const metadata: Metadata = {
   title: "LoL 도감",
@@ -17,17 +16,10 @@ export default function RootLayout({
   return (
     <html lang="ko-kr">
       <body>
-        <ThemeProvider
-          attribute="class"
-          defaultTheme="dark"
-          enableSystem
-          disableTransitionOnChange
-        >
-          <div className="min-h-screen bg-gradient-to-b from-background to-background/90">
-            <Header />
-            <main className="container mx-auto px-4 py-6">{children}</main>
-          </div>
-        </ThemeProvider>
+        <div className="min-h-screen bg-gradient-to-b from-background to-background/90">
+          <Header />
+          <main className="container mx-auto px-4 py-6">{children}</main>
+        </div>
       </body>
     </html>
   );
